moving all tests out of /test folders (easier to see where the coverage is while browsing project)
