+++ src/components/containers/Slideshow/children/SlideshowSettingsButton/SlideshowSettingsButton.spec.js
@@ -4,7 +4,7 @@ import { mount } from 'enzyme';
 import { mockStore } from 'testUtils';
 import { ACTIONS } from 'constants';
 
-import SlideshowSettingsButton from '../SlideshowSettingsButton';
+import SlideshowSettingsButton from './SlideshowSettingsButton';
 
 describe('SlideshowSettingsButton', () => {
     it(`Should dispatch ${ACTIONS.SLIDESHOW_SETTINGS_TOGGLE} when rendered and clicked`, (done) => {
